Made an abstract base class for all stack implementations to inherit from. This reduces boilerplate code a lot and limits the requirements of what the concrete classes implements to just the logic that pertains to their type of stack.
